pyproject.toml
[project]
name = "python-project-83"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "flask>=3.1.1",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.0",
    "validators>=0.35.0",
]

[project.scripts]
page-analyzer = "page_analyzer:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["page_analyzer"]

[dependency-groups]
dev = [
    "ruff>=0.11.10",
]
